Got Rid of Bold Export -The function used to bold company names is no longer exported from Mention. Instead It is passed down from DashboardBody to Mention and Dialog
